+++ tests/test_main.py
@@ -3,12 +3,24 @@
 
 import unittest
 import sys
+import os
+import stat
+import shutil
 from pathlib import Path
+
+from . import test_dir
 
 import pkg_about
 
 
 class MainTestCase(unittest.TestCase):
+
+    @staticmethod
+    def is_namedtuple_instance(obj):
+        return (isinstance(obj, tuple) and
+                hasattr(obj, '_fields') and
+                isinstance(obj._fields, tuple) and
+                all(isinstance(field, str) for field in obj._fields))
 
     @classmethod
     def setUpClass(cls):
@@ -26,39 +38,96 @@ class MainTestCase(unittest.TestCase):
         cls.version_micro_expected = int(version_parts[2])
 
     def test_about(self):
-        pkg_about.about("pkg_about")
-        self.assertEqual(__title__, "pkg_about")
-        self.assertEqual(__version__, self.version_expected)
-        self.assertEqual(__version_info__.major, self.version_major_expected)
-        self.assertEqual(__version_info__.minor, self.version_minor_expected)
-        self.assertEqual(__version_info__.micro, self.version_micro_expected)
-        self.assertEqual(__version_info__.releaselevel, "final")
-        self.assertEqual(__version_info__.serial, 0)
-        self.assertEqual(__summary__, "Shares Python package metadata at runtime.")
-        self.assertEqual(__uri__, "https://pypi.org/project/pkg-about/")
-        self.assertEqual(__author__, "Adam Karpierz")
-        self.assertEqual(__email__, "adam@karpierz.net")
-        self.assertEqual(__author_email__, "adam@karpierz.net")
-        self.assertEqual(__maintainer__, "Adam Karpierz")
-        self.assertEqual(__maintainer_email__, "adam@karpierz.net")
-        self.assertEqual(__license__, "Zlib")
-        self.assertEqual(__copyright__, __author__)
-
-    def test_about_from_setup(self):
-        pkg_about.about_from_setup(Path(__file__).resolve().parent.parent)
+        about = pkg_about.about("pkg_about")
+        self.assertIsInstance(about, dict)
+        self.assertIs(__title__, about.__title__)
         self.assertEqual(about.__title__, "pkg_about")
+        self.assertIs(__version__, about.__version__)
         self.assertEqual(about.__version__, self.version_expected)
+        self.assertIs(__version_info__, about.__version_info__)
+        self.assertTrue(self.is_namedtuple_instance(about.__version_info__))
         self.assertEqual(about.__version_info__.major, self.version_major_expected)
         self.assertEqual(about.__version_info__.minor, self.version_minor_expected)
         self.assertEqual(about.__version_info__.micro, self.version_micro_expected)
         self.assertEqual(about.__version_info__.releaselevel, "final")
         self.assertEqual(about.__version_info__.serial, 0)
+        self.assertIs(__summary__, about.__summary__)
         self.assertEqual(about.__summary__, "Shares Python package metadata at runtime.")
+        self.assertIs(__uri__, about.__uri__)
         self.assertEqual(about.__uri__, "https://pypi.org/project/pkg-about/")
+        self.assertIs(__author__, about.__author__)
         self.assertEqual(about.__author__, "Adam Karpierz")
-        self.assertEqual(about.__email__, "adam@karpierz.net")
-        self.assertEqual(about.__author_email__, "adam@karpierz.net")
+        self.assertIs(__email__, about.__email__)
+        self.assertEqual(about.__email__, about.__author_email__)
+        self.assertIs(__author_email__, about.__author_email__)
+        self.assertEqual(about.__author_email__, "Adam Karpierz <adam@karpierz.net>")
+        self.assertIs(__maintainer__, about.__maintainer__)
         self.assertEqual(about.__maintainer__, "Adam Karpierz")
-        self.assertEqual(about.__maintainer_email__, "adam@karpierz.net")
+        self.assertIs(__maintainer_email__, about.__maintainer_email__)
+        self.assertEqual(about.__maintainer_email__, "Adam Karpierz <adam@karpierz.net>")
+        self.assertIs(__license__, about.__license__)
         self.assertEqual(about.__license__, "Zlib")
-        self.assertEqual(about.__copyright__, about.__author__)
+        self.assertIs(__copyright__, about.__copyright__)
+        self.assertEqual(about.__copyright__, __author__)
+
+    def test_about_from_setup(self):
+        package_path = Path(__file__).resolve().parent.parent
+        setup_cfg = shutil.copy(test_dir/"data/setup.cfg", package_path/"setup.cfg")
+        try:
+            ret_about = pkg_about.about_from_setup(package_path)
+            self.assertIs(ret_about, about)
+            self.assertIsInstance(about, dict)
+            self.assertEqual(about.__title__, "pkg_about")
+            self.assertEqual(about.__version__, self.version_expected)
+            self.assertTrue(self.is_namedtuple_instance(about.__version_info__))
+            self.assertEqual(about.__version_info__.major, self.version_major_expected)
+            self.assertEqual(about.__version_info__.minor, self.version_minor_expected)
+            self.assertEqual(about.__version_info__.micro, self.version_micro_expected)
+            self.assertEqual(about.__version_info__.releaselevel, "final")
+            self.assertEqual(about.__version_info__.serial, 0)
+            self.assertEqual(about.__summary__, "Shares Python package metadata at runtime.")
+            self.assertEqual(about.__uri__, "https://pypi.org/project/pkg-about/")
+            self.assertEqual(about.__author__, "Adam Karpierz")
+            self.assertEqual(about.__email__, about.__author_email__)
+            self.assertEqual(about.__author_email__, "Adam Karpierz <adam@karpierz.net>")
+            self.assertEqual(about.__maintainer__, "Adam Karpierz")
+            self.assertEqual(about.__maintainer_email__, "Adam Karpierz <adam@karpierz.net>")
+            self.assertEqual(about.__license__, "Zlib")
+            self.assertEqual(about.__copyright__, about.__author__)
+        finally:
+            os.chmod(setup_cfg, stat.S_IWRITE)
+            setup_cfg.unlink()
+
+    def test_about_appdirs(self):
+        about = pkg_about.about("appdirs")
+        self.assertIsInstance(about, dict)
+        self.assertIs(__title__, about.__title__)
+        self.assertEqual(about.__title__, "appdirs")
+        self.assertIs(__version__, about.__version__)
+        self.assertEqual(about.__version__, "1.4.4")
+        self.assertIs(__version_info__, about.__version_info__)
+        self.assertTrue(self.is_namedtuple_instance(about.__version_info__))
+        self.assertEqual(about.__version_info__.major, 1)
+        self.assertEqual(about.__version_info__.minor, 4)
+        self.assertEqual(about.__version_info__.micro, 4)
+        self.assertEqual(about.__version_info__.releaselevel, "final")
+        self.assertEqual(about.__version_info__.serial, 0)
+        self.assertIs(__summary__, about.__summary__)
+        self.assertEqual(about.__summary__, ('A small Python module for determining appropriate'
+                                             ' platform-specific dirs, e.g. a "user data dir".'))
+        self.assertIs(__uri__, about.__uri__)
+        self.assertEqual(about.__uri__, "http://github.com/ActiveState/appdirs")
+        self.assertIs(__author__, about.__author__)
+        self.assertEqual(about.__author__, "Trent Mick")
+        self.assertIs(__email__, about.__email__)
+        self.assertEqual(about.__email__, about.__author_email__)
+        self.assertIs(__author_email__, about.__author_email__)
+        self.assertEqual(about.__author_email__, "trentm@gmail.com")
+        self.assertIs(__maintainer__, about.__maintainer__)
+        self.assertEqual(about.__maintainer__, "Jeff Rouse")
+        self.assertIs(__maintainer_email__, about.__maintainer_email__)
+        self.assertEqual(about.__maintainer_email__, "jr@its.to")
+        self.assertIs(__license__, about.__license__)
+        self.assertEqual(about.__license__, "MIT")
+        self.assertIs(__copyright__, about.__copyright__)
+        self.assertEqual(about.__copyright__, __author__)